students_folder/Nikulina/lab7/requirements.txt
fastapi==0.110.2
pydantic==2.7.0
SQLAlchemy==2.0.30
uvicorn==0.29.0
asyncpg==0.29.0
pytest
pytest-asyncio
httpx
pyyaml